chore: set CLI as first project in solution
Because Visual Studio stores the startup project in the user settings for a solution, when one is not set up yet -- such as after cloning the repo locally and running it for the first time -- it will default to the first project listed in the .sln file.
Since the CLI project is the end goal, switching up the order in the file to fit this behavior.

#100